backend/pyproject.toml
[tool.poetry]
name = "backend"
version = "0.1.0"
description = ""
authors = ["Your Name <you@example.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = ">=3.11.4"
django = ">=4.2.4"
django-cors-headers = "^4.2.0"
djangorestframework = ">=3.14.0"
djangorestframework-simplejwt = {extras = ["crypto"], version = "^5.2.2"}
whitenoise = ">=6.5.0"

[tool.poetry.group.dev.dependencies]
pytest = ">=7.4.0"
pytest-django = ">=4.5.2"
black = ">=23.3.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
